Dash in all frames fixed
Now it should have the same speed in all framerates

+++ Forest Dash/Motor2D/j1Player.h
@@ -107,7 +107,7 @@ private:
 	int gravity;
 	int speedX;
 	int speedY;
-	int dashspeed;
+	float dashspeed;
 	float acceleration;
 	float speedBar;
